Relative links restructuring
