+++ src/Components/Pages/Team/Ancilliary/Components/Card.js
@@ -1,8 +1,9 @@
 import React from 'react'
-import defaultProfileImage from '../../../../Resources/Images/defaultProfileImage.png';
-import { Icon } from '../../../Shared/Icon';
-import { githubIcon, linkedinIcon, instagramIcon } from "../../../../Resources/Icons";
-import { ICON_DEFAULT_COLOR } from '../../../../Constants';
+
+import {Icon} from '../../../../Shared/Icon'
+import { linkedinIcon, githubIcon, instagramIcon } from '../../../../../Resources/Icons'
+import {ICON_DEFAULT_COLOR} from '../../../../../Constants'
+import defaultProfileImage from '../../../../../Resources/Images/defaultProfileImage.png'
 
 const linkToIcon = {
   github: githubIcon,
