+++ src/Components/Pages/Team/index.js
@@ -2,7 +2,7 @@ import React, { useContext } from 'react';
 import { teamIcon } from '../../../Resources/Icons';
 import Title, { Subtitle } from '../../Shared/Title';
 import Wrapper from '../../Shared/Wrapper';
-import { CurrentTeamCard, PastMembersCard } from './Components/Card';
+import { CurrentTeamCard, PastMembersCard } from './Ancilliary/Components/Card';
 import { currentTeam, pastMembers } from './Ancilliary/Data/data';
 import { NavbarToggleContext } from '../../../Context/navbarToggle';
 import Navbar from '../../Shared/Navbar';
