+++ src/Components/Pages/Visualizer/Ancilliary/Computation/MVT.js
@@ -1,6 +1,6 @@
 // MVT GRAPH COMPUTATION 
 
-import { COLORS_MVT_GRAPH } from "../../../../Constants";
+import { COLORS_MVT_GRAPH } from "../../../../../Constants";
 
 
 export const mvtGraphComputation = (transactionData, labelToggle, chartMaxDataUpdate, messageChartDataUpdate) => {
